refactor for serialization module
